Move ncurses_wrapper into its own module
