+++ src/fftw/ext.rs
@@ -1,4 +1,3 @@
-
 extern crate libc;
 use self::libc::{c_int, size_t, c_void};
 use fftw::types::*;
