+++ src/ncurses/mod.rs
@@ -1,0 +1,2 @@
+mod ext;
+pub mod window;
