+++ src/ncurses/window.rs
@@ -1,9 +1,9 @@
 #![allow(missing_copy_implementations)]
-
 extern crate libc;
 
 use std::ffi::CString;
 use self::libc::{c_int, c_char};
+use ncurses::ext;
 
 
 /// Safe wrapper for the ncurses endwin function. Call this when you are done
@@ -86,28 +86,5 @@ fn handle_err(result: c_int) -> Result<c_int, c_int> {
         Err(result)
     } else {
         Ok(result)
-    }
-}
-
-
-mod ext {
-    /// Module for external ncurses functions and types
-
-    extern crate libc;
-    use self::libc::{c_int, c_char};
-
-    #[repr(C)]
-    pub struct WINDOW;
-
-    #[link(name="ncurses")]
-    extern {
-        pub fn initscr() -> *mut WINDOW;
-        pub fn endwin() -> c_int;
-        pub fn wrefresh(win: *mut WINDOW) -> c_int;
-        pub fn mvwaddstr(win: *mut WINDOW, y: c_int, x: c_int, text: *const c_char) -> c_int;
-        pub fn mvwaddnstr(win: *mut WINDOW, y: c_int, x: c_int, text: *const c_char, n: c_int) -> c_int;
-        pub fn getmaxy(win: *mut WINDOW) -> c_int;
-        pub fn getmaxx(win: *mut WINDOW) -> c_int;
-        pub fn curs_set(visibility: c_int) -> c_int;
     }
 }
